requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.3
gast==0.3.3
get==0.0.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
h5py==2.10.0
httplib2==0.18.1
idna==2.9
Keras-Preprocessing==1.1.2
Markdown==3.2.2
multidict==4.7.6
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
post==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
query-string==0.0.0
request==0.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
